+++ src/Benchmarking/DistanceMatrixBenchmark.cpp
@@ -11,7 +11,7 @@ double DistanceMatrixBenchmark::benchmark(const vector < pair < unsigned int, un
     Timer dijkstraTimer("Distance Matrix trips benchmark");
     dijkstraTimer.begin();
 
-    for(unsigned int i = 0; i < trips.size(); i++) {
+    for(size_t i = 0; i < trips.size(); i++) {
         distances[i] = matrix.findDistance(trips.at(i).first, trips.at(i).second);
     }
