+++ src/Benchmarking/TNRAFBenchmark.cpp
@@ -15,7 +15,7 @@ double TNRAFBenchmark::benchmark(const vector < pair < unsigned int, unsigned in
     Timer tnrTimer("Transit Node Routing with Arc Flags trips benchmark");
     tnrTimer.begin();
 
-    for(unsigned int i = 0; i < trips.size(); i++) {
+    for(size_t i = 0; i < trips.size(); i++) {
         distances[i] = queryManager.findDistance(trips.at(i).first, trips.at(i).second);
     }
 
@@ -30,7 +30,7 @@ double TNRAFBenchmark::benchmarkWithMapping(const vector < pair < long long unsi
     Timer tnrTimer("Transit Node Routing with Arc Flags trips benchmark with mapping");
     tnrTimer.begin();
 
-    for(unsigned int i = 0; i < trips.size(); i++) {
+    for(size_t i = 0; i < trips.size(); i++) {
         distances[i] = queryManager.findDistance(trips.at(i).first, trips.at(i).second);
     }
