+++ src/Benchmarking/TNRBenchmark.cpp
@@ -15,7 +15,7 @@ double TNRBenchmark::benchmark(const vector < pair < unsigned int, unsigned int>
     Timer tnrTimer("Transit Node Routing trips benchmark");
     tnrTimer.begin();
 
-    for(unsigned int i = 0; i < trips.size(); i++) {
+    for(size_t i = 0; i < trips.size(); i++) {
         distances[i] = queryManager.findDistance(trips.at(i).first, trips.at(i).second);
     }
 
@@ -30,7 +30,7 @@ double TNRBenchmark::benchmarkWithMapping(const vector < pair < long long unsign
     Timer tnrTimer("Transit Node Routing trips benchmark with mapping");
     tnrTimer.begin();
 
-    for(unsigned int i = 0; i < trips.size(); i++) {
+    for(size_t i = 0; i < trips.size(); i++) {
         distances[i] = queryManager.findDistance(trips.at(i).first, trips.at(i).second);
     }
