+++ src/DistanceMatrix/DistanceMatrixCsvOutputter.cpp
@@ -19,7 +19,7 @@ void DistanceMatrixCsvOutputter::store(DistanceMatrix &dm, const string &path) {
 
     const vector<dist_t> &distances = dm.getRawData();
     const auto nodesCnt = dm.nodes();
-    for (unsigned int i = 0; i < distances.size(); i++) {
+    for(size_t i = 0; i < distances.size(); i++) {
         output << distances[i];
 
         if ((i + 1) % nodesCnt == 0) {
